feat: Add unique device identity for userId and xmtpId (#69)
* add unique device identity for userId and xmtpId

* fix

* fix

* fix identities routes and tests

* cleaner tests
