+++ .cursorrules
@@ -34,6 +34,11 @@ When asked to enter "Planner Mode" deeply reflect upon the changes being asked a
 - Use dedicated handlers directory for route logic
 - Follow RESTful principles for endpoint naming
 - Implement consistent error handling using AppError class
+- Always place the `return;` statement on a new line immediately after sending a response with `res.status().json()` or similar methods. For example:
+  ```typescript
+  res.status(400).json({ error: "Invalid request body" });
+  return;
+  ```
 
 ### General
