+++ tests/identities.test.ts
@@ -9,17 +9,19 @@ import {
   test,
 } from "bun:test";
 import express from "express";
-import identitiesRouter from "@/api/v1/identities";
+import identitiesRouter from "@/api/v1/identities/identities.router";
 import { jsonMiddleware } from "@/middleware/json";
 import { prisma } from "@/utils/prisma";
 
 const app = express();
 app.use(jsonMiddleware);
 
+const AUTH_USER_XMTP_ID = "test-xmtp-id";
+
 // Add middleware to simulate authentication for tests
 app.use((req, res, next) => {
   // Set xmtpId for testing - this simulates the auth middleware
-  req.app.locals.xmtpId = "test-xmtp-id";
+  req.app.locals.xmtpId = AUTH_USER_XMTP_ID;
   next();
 });
 
@@ -48,29 +50,23 @@ beforeEach(async () => {
 describe("/identities API", () => {
   const testDeviceId = "test-device-id";
   const testUserId = "test-user-id";
+  const testDeviceName = "Test Device";
+  const testDeviceOs = "ios";
+  const testTurnkeyUserId = "test-identities-turnkey-user-id";
 
   beforeEach(async () => {
-    // create a test user and device before each test
+    // create a device and a user
     await prisma.device.create({
       data: {
         id: testDeviceId,
-        name: "Test Device",
-        os: "ios",
+        name: testDeviceName,
+        os: testDeviceOs,
         user: {
           create: {
-            id: "test-user-id",
-            turnkeyUserId: "test-identities-turnkey-user-id",
+            id: testUserId,
+            turnkeyUserId: testTurnkeyUserId,
           },
         },
-      },
-    });
-
-    // Add a device identity with the same xmtpId as in the middleware
-    await prisma.deviceIdentity.create({
-      data: {
-        userId: testUserId,
-        turnkeyAddress: "test-turnkey-address",
-        xmtpId: "test-xmtp-id",
       },
     });
   });
@@ -85,7 +81,7 @@ describe("/identities API", () => {
         },
         body: JSON.stringify({
           turnkeyAddress: "0x123",
-          xmtpId: "test-xmtp-id",
+          xmtpId: AUTH_USER_XMTP_ID,
         }),
       },
     );
@@ -93,7 +89,7 @@ describe("/identities API", () => {
 
     expect(response.status).toBe(201);
     expect(identity.turnkeyAddress).toBe("0x123");
-    expect(identity.xmtpId).toBe("test-xmtp-id");
+    expect(identity.xmtpId).toBe(AUTH_USER_XMTP_ID);
 
     const response2 = await fetch(
       `http://localhost:3003/identities/device/invalid-device-id`,
@@ -104,7 +100,7 @@ describe("/identities API", () => {
         },
         body: JSON.stringify({
           turnkeyAddress: "0x123",
-          xmtpId: "test-xmtp-id",
+          xmtpId: AUTH_USER_XMTP_ID,
         }),
       },
     );
@@ -125,7 +121,7 @@ describe("/identities API", () => {
         },
         body: JSON.stringify({
           turnkeyAddress: "0x123",
-          xmtpId: "test-xmtp-id",
+          xmtpId: AUTH_USER_XMTP_ID,
         }),
       },
     );
@@ -153,7 +149,7 @@ describe("/identities API", () => {
         },
         body: JSON.stringify({
           turnkeyAddress: "0x123",
-          xmtpId: "test-xmtp-id",
+          xmtpId: AUTH_USER_XMTP_ID,
         }),
       },
     );
@@ -167,7 +163,7 @@ describe("/identities API", () => {
     expect(response.status).toBe(200);
     expect(identity.id).toBe(createdIdentity.id);
     expect(identity.turnkeyAddress).toBe("0x123");
-    expect(identity.xmtpId).toBe("test-xmtp-id");
+    expect(identity.xmtpId).toBe(AUTH_USER_XMTP_ID);
   });
 
   test("PUT /identities/:identityId updates identity", async () => {
@@ -181,7 +177,7 @@ describe("/identities API", () => {
         },
         body: JSON.stringify({
           turnkeyAddress: "0x123",
-          xmtpId: "test-xmtp-id",
+          xmtpId: AUTH_USER_XMTP_ID,
         }),
       },
     );
@@ -219,7 +215,7 @@ describe("/identities API", () => {
         },
         body: JSON.stringify({
           turnkeyAddress: "0x123",
-          xmtpId: "test-xmtp-id",
+          xmtpId: AUTH_USER_XMTP_ID,
         }),
       },
     );
@@ -283,7 +279,7 @@ describe("/identities API", () => {
         },
         body: JSON.stringify({
           turnkeyAddress: "0x123",
-          xmtpId: "test-xmtp-id",
+          xmtpId: AUTH_USER_XMTP_ID,
         }),
       },
     );
@@ -325,7 +321,7 @@ describe("/identities API", () => {
         },
         body: JSON.stringify({
           // missing required turnkeyAddress
-          xmtpId: "test-xmtp-id",
+          xmtpId: AUTH_USER_XMTP_ID,
         }),
       },
     );
@@ -346,7 +342,7 @@ describe("/identities API", () => {
         },
         body: JSON.stringify({
           turnkeyAddress: "0x123",
-          xmtpId: "test-xmtp-id",
+          xmtpId: AUTH_USER_XMTP_ID,
         }),
       },
     );
@@ -358,8 +354,7 @@ describe("/identities API", () => {
     const identities = (await response.json()) as DeviceIdentity[];
 
     expect(response.status).toBe(200);
-    // Expect 2 identities because we're creating one in the setup and one in this test
-    expect(identities).toHaveLength(2);
+    expect(identities).toHaveLength(1);
 
     // Verify the identity we just created is in the list
     const foundIdentity = identities.find(
@@ -391,9 +386,8 @@ describe("/identities API", () => {
     const identities2 = (await response2.json()) as DeviceIdentity[];
 
     expect(response2.status).toBe(200);
-    // We still have 2 identities after unlinking because unlinking only removes the
-    // identity-device relationship, not the identity itself
-    expect(identities2).toHaveLength(2);
+    // We expect 1 identity because we unlinked the identity from the device but we still have the identity in the database
+    expect(identities2).toHaveLength(1);
 
     const foundIdentityAfterUnlink = identities2.find(
       (identity) => identity.id === createdIdentity.id,
@@ -409,9 +403,7 @@ describe("/identities API", () => {
       `http://localhost:3003/identities/user/${nonExistentUserId}`,
     );
 
-    // Now we expect a 403 since the user doesn't have a device identity with the auth xmtpId
-    expect(response.status).toBe(403);
-    const data = (await response.json()) as { error: string };
-    expect(data.error).toBe("Not authorized to access this user's identities");
+    const identities = (await response.json()) as DeviceIdentity[];
+    expect(identities).toHaveLength(0);
   });
 });
